Safari-609.1.11.1/LayoutTests/fast/dom/Range/range-clone-empty.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
description(
"This test checks cloning an empty range returns an empty fragment."
);

var r = document.createRange();
shouldBeTrue("r.cloneContents() != undefined");
shouldBeTrue("r.cloneContents() != null");
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
