third_party/blink/web_tests/platform/generic/fast/dom/Range/range-clone-empty-expected.txt

This test checks cloning an empty range returns an empty fragment.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS r.cloneContents() != undefined is true
PASS r.cloneContents() != null is true
PASS successfullyParsed is true

TEST COMPLETE
